New changes

+++ app/Http/Controllers/NotesController.php
@@ -84,6 +84,12 @@ class NotesController extends Controller
 
     public function save_ticket(Request $request, $id)
     {
+
+        $request->validate([
+            'payment' => 'required|between:0,99999.99',
+            'payment_method' => 'required|string|max:20',
+        ]);
+
         $total = 0;
         $mesa = Mesa::where('id',$id)->first();
         $waiter = User::select('id','name')
@@ -95,27 +101,68 @@ class NotesController extends Controller
                 ->join('active_products','active_products.product_id','=','products.id')
                 ->where('active_products.active_id',$id)
                 ->get();
-        
-        foreach($products as $item)
-        {
-            $amount_price = $item->amount * $item->price;
-            $item->setAttribute('amount_price',$amount_price);
-            $total = $total + $amount_price; 
+        if(isset($products)){
+            foreach($products as $item)
+            {
+                $amount_price = $item->amount * $item->price;
+                $item->setAttribute('amount_price',$amount_price);
+                $total = $total + $amount_price; 
+            }
+
+            if($request->payment_method == "Tarjeta")
+            {
+                $mesa->setAttribute('waiter',$waiter);
+                $mesa->setAttribute('consumes',$products);
+                $mesa->setAttribute('total',$total);
+
+                $mesa->setAttribute('payment_method',$request->payment_method);
+                $mesa->setAttribute('payment',$total);
+                $mesa->setAttribute('change',0);
+
+                
+                $ticket = new Ticket;
+                $ticket->purchase_info = $mesa;
+                $ticket->save();
+
+                $active = ActiveTables::where('mesa_id',$id)->delete();
+
+                return response()->json(['message' => 'Se guardó correctamente la compra']);
+            }
+            else if($request->payment_method == "Efectivo") {
+
+                if($request->payment >= $total)
+                {
+                    $change = $request->payment -$total;
+
+                    $mesa->setAttribute('waiter',$waiter);
+                    $mesa->setAttribute('consumes',$products);
+                    $mesa->setAttribute('total',$total);
+
+                    $mesa->setAttribute('payment_method',$request->payment_method);
+                    $mesa->setAttribute('payment',$request->payment);
+                    $mesa->setAttribute('change',$change);
+
+                    
+                    $ticket = new Ticket;
+                    $ticket->purchase_info = $mesa;
+                    $ticket->save();
+
+                    $active = ActiveTables::where('mesa_id',$id)->delete();
+
+                    return response()->json(['message' => 'Se guardó correctamente la compra']);
+                }
+                else {
+                    return response()->json(['message' => 'El pago no es sufiente para pagar el ticket']);
+                }
+            }
+            else {
+                return response()->json(['message' => 'No especifico el metodo de pago']);
+            }
         }
-        $mesa->setAttribute('waiter',$waiter);
-        $mesa->setAttribute('consumes',$products);
-        $mesa->setAttribute('total',$total);
-
-        $mesa->setAttribute('payment',$request->payment);
-        $mesa->setAttribute('change',$request->change);
-
-        $ticket = new Ticket;
-        $ticket->purchase_info = $mesa;
-        $ticket->save();
-
-        $active = ActiveTables::where('mesa_id',$id)->delete();
+        else {
+            return response()->json(['message' => 'La mesa no tiene productos']);
+        }
         
-        return response()->json(['message' => 'Se guardó correctamente la compra']);
     }
 
 }
